requirements.txt
astroid==1.6.0
backports.functools-lru-cache==1.4
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
enum34==1.1.6
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
pylint==1.8.1
requests==2.18.4
selenium==3.8.1
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
wrapt==1.10.11
